update fareProductTypeConfig version for onBehalfOfEnabled
