+++ package.json
@@ -27,7 +27,7 @@
   },
   "dependencies": {
     "@adrianso/react-native-device-brightness": "^1.2.7",
-    "@atb-as/config-specs": "^3.17.0",
+    "@atb-as/config-specs": "^3.18.0",
     "@atb-as/generate-assets": "^10.0.0",
     "@atb-as/theme": "^8.0.0",
     "@bugsnag/react-native": "^7.21.0",
