+++ src/stacks-hierarchy/Root_PurchaseOverviewScreen/components/TravellerSelectionSheet.tsx
@@ -1,6 +1,9 @@
 import {BottomSheetContainer} from '@atb/components/bottom-sheet';
 import {PurchaseOverviewTexts, useTranslation} from '@atb/translations';
-import {TravellerSelectionMode} from '@atb/configuration';
+import {
+  FareProductTypeConfig,
+  TravellerSelectionMode,
+} from '@atb/configuration';
 import {ScrollView} from 'react-native';
 import React, {useState} from 'react';
 import {StyleSheet} from '@atb/theme';
@@ -14,8 +17,7 @@ import {UserProfileWithCount} from '@atb/fare-contracts';
 
 type TravellerSelectionSheetProps = {
   selectionMode: TravellerSelectionMode;
-  fareProductType: string;
-  fareProductOnBehalfOfEnabled: boolean;
+  fareProductTypeConfig: FareProductTypeConfig;
   selectableUserProfilesWithCountInit: UserProfileWithCount[];
   onConfirmSelection: (
     chosenSelectableUserProfiles: UserProfileWithCount[],
@@ -25,8 +27,7 @@ type TravellerSelectionSheetProps = {
 };
 export const TravellerSelectionSheet = ({
   selectionMode,
-  fareProductType,
-  fareProductOnBehalfOfEnabled,
+  fareProductTypeConfig,
   selectableUserProfilesWithCountInit,
   onConfirmSelection,
   isOnBehalfOfToggle,
@@ -54,8 +55,7 @@ export const TravellerSelectionSheet = ({
       <ScrollView style={style.container}>
         {selectionMode === 'multiple' ? (
           <MultipleTravellersSelection
-            fareProductType={fareProductType}
-            fareProductOnBehalfOfEnabled={fareProductOnBehalfOfEnabled}
+            fareProductTypeConfig={fareProductTypeConfig}
             {...userCountState}
             userProfilesWithCount={selectableUserProfilesWithCount}
             setIsTravelerOnBehalfOfToggle={setIsTravelerOnBehalfOfToggle}
@@ -63,8 +63,7 @@ export const TravellerSelectionSheet = ({
           />
         ) : (
           <SingleTravellerSelection
-            fareProductType={fareProductType}
-            fareProductOnBehalfOfEnabled={fareProductOnBehalfOfEnabled}
+            fareProductTypeConfig={fareProductTypeConfig}
             {...userCountState}
             userProfilesWithCount={selectableUserProfilesWithCount}
             setIsTravelerOnBehalfOfToggle={setIsTravelerOnBehalfOfToggle}
