+++ src/stacks-hierarchy/Root_PurchaseOverviewScreen/components/Travellers/SingleTravellerSelection.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import {UserCountState} from './use-user-count-state';
 import {
   useTranslation,
   TicketTravellerTexts,
@@ -17,22 +16,22 @@ import {View} from 'react-native';
 import {StyleSheet} from '@atb/theme';
 import {HoldingHands} from '@atb/assets/svg/color/images';
 import {useOnBehalfOf} from '@atb/on-behalf-of';
-import {TravelerOnBehalfOfProps} from './types';
+import {TravellerSelectionBottomSheetType} from './types';
 
 export function SingleTravellerSelection({
   userProfilesWithCount,
   addCount,
   removeCount,
-  fareProductType,
-  fareProductOnBehalfOfEnabled,
+  fareProductTypeConfig,
   setIsTravelerOnBehalfOfToggle,
   isTravelerOnBehalfOfToggle,
-}: UserCountState & TravelerOnBehalfOfProps) {
+}: TravellerSelectionBottomSheetType) {
   const {t, language} = useTranslation();
   const styles = useStyles();
   const selectedProfile = userProfilesWithCount.find((u) => u.count);
 
-  const isOnBehalfOfEnabled = useOnBehalfOf() && fareProductOnBehalfOfEnabled;
+  const isOnBehalfOfEnabled =
+    useOnBehalfOf() && fareProductTypeConfig.configuration.onBehalfOfEnabled;
 
   const select = (u: UserProfileWithCount) => {
     if (selectedProfile) {
@@ -68,7 +67,7 @@ export function SingleTravellerSelection({
         keyExtractor={(u) => u.userTypeString}
         itemToText={(u) => getReferenceDataName(u, language)}
         itemToSubtext={(u) =>
-          travellerInfoByFareProductType(fareProductType, u)
+          travellerInfoByFareProductType(fareProductTypeConfig.type, u)
         }
         selected={selectedProfile}
         onSelect={select}
